refactor: move HalfCheetah compile benchmark script
